+++ src/frontend/components/routes/utils/wrapper.tsx
@@ -15,6 +15,11 @@ const StyledWrapper = styled(Box)`
   }
 `
 
+StyledWrapper.defaultProps = {
+  // width: [1, 1, 1, 1024],
+  // m: '0 auto',
+}
+
 const Wrapper: React.FC<BoxProps> = (props) => {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const { children, variant, color, ...rest } = props