+++ src/frontend/components/spec/action-json.factory.ts
@@ -8,4 +8,5 @@ factory.define<ActionJSON>('ActionJSON', Object, {
   label: 'someLabel',
   showFilter: false,
   resourceId: 'resource',
+  hideActionHeader: false,
 })